requirement.txt
accelerate==0.16.0
einops
ema-pytorch
pytorch-lightning==1.9.3
scikit-learn
scipy
thop
timm==0.6.12
tensorboard
fvcore
albumentations
omegaconf
numpy==1.23.5